+++ fail/fail.go
@@ -46,6 +46,7 @@ const (
 	ErrDotOperatorNotSupported = "the dot operator is not supported for type '%s'"
 	ErrPropertyNotFound        = "property '%s' not found in type '%s'"
 	ErrDivisionByZero          = "division by zero error. The right-hand side of the division operator must not be zero"
+	ErrSomeDirsOnlyInTemplates = "you cannot use @use, @insert, @reserve and @component in simple file or string execution, only in templates"
 
 	// Functions
 	ErrNoFuncForThisType  = "function '%s' doesn't exist for type '%s'"